+++ src/attributes.js
@@ -1,0 +1,10 @@
+export const footnoteAttributes = {
+    reference: {
+        type: 'string',
+        default: '',
+    },
+    referenceId: {
+        type: 'string',
+        default: '',
+    }
+};